+++ src/root.zig
@@ -1,4 +1,5 @@
 const zua = @import("zua");
 
 pub const Lumem = @import("lumem/lumem.zig");
+pub const PinWatcher = @import("lumem/mem/pin.zig").PinWatcher;
 const e = @import("lumem/mem/entry.zig");